digital_profile/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==20.1.0
black==19.10b0
click==7.1.2
Django==3.1
djangorestframework==3.11.1
pathspec==0.8.0
Pillow==7.2.0
pytz==2020.1
regex==2020.7.14
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
